WEB: Map Manager now ask for confirmation when deleting a map.
git-svn-id: https://itsd.ict.sait.ca:9880/capstone07/trunk@255 0c6396e7-dfa7-ad41-aac3-ea9b0ac9ed18

+++ web/webapp/js/MapController.js
@@ -271,17 +271,22 @@ var MapController = function() {
         
         delMap: function(onDeleteMap) {
         	if (currentMap !== null) {
-        		var that = this;
-    			$.ajax({
-    				type : "DELETE",
-    				url : ctx + "/api/map/" + selectedMapId,
-    				contentType : "application/json; charset=utf-8",
-    				complete : ajaxMessageClosureOnError(messagable, function(ok, json) {
-    					showOkMessage(messagable, "Map deleted");
-    					clearMap();
-    					onDeleteMap();
-    				})
-    			});
+        		var messagable = createElemMessagable("#message");
+
+        		var r = confirm("Are you sure you want to delete this map");
+        		if(r === true){
+        			var that = this;
+        			$.ajax({
+        				type : "DELETE",
+        				url : ctx + "/api/map/" + selectedMapId,
+        				contentType : "application/json; charset=utf-8",
+        				complete : ajaxMessageClosureOnError(messagable, function(ok, json) {
+        					showOkMessage(messagable, "Map deleted");
+        					clearMap();
+        					onDeleteMap();
+        				})
+        			});
+        		}
             }
         },
